pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ado-ai-cli"
version = "0.1.0"
description = "Azure DevOps AI Auto-Complete CLI Tool powered by Claude"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Spine Zhang"}
]
keywords = ["azure-devops", "cli", "ai", "claude", "automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # CLI & Output
    "typer[all]>=0.12.0",
    "rich>=13.7.0",

    # Azure DevOps
    "azure-devops>=7.1.0b4",

    # AI Integration
    "anthropic>=0.34.0",

    # Configuration
    "python-dotenv>=1.0.0",
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",

    # Utilities
    "tenacity>=9.0.0",

    # Web Framework (Phase 1: Foundation)
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "jinja2>=3.1.3",
    "sqlalchemy>=2.0.25",
    "alembic>=1.13.0",
    "cryptography>=41.0.7",
    "python-multipart>=0.0.6",
    "websockets>=12.0",
    "slowapi>=0.1.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.0",
    "black>=24.8.0",
    "ruff>=0.5.0",
    "mypy>=1.11.0",
]

[project.scripts]
ado-ai = "ado_ai_cli.cli:app"

[project.urls]
Homepage = "https://github.com/yourusername/ado-ai-cli"
Repository = "https://github.com/yourusername/ado-ai-cli"
Issues = "https://github.com/yourusername/ado-ai-cli/issues"

[tool.setuptools]
packages = ["ado_ai_cli"]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
ado_ai_cli = ["py.typed"]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.mypy]
python_version = "3.9"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=ado_ai_cli --cov-report=term-missing"
